Revert "make all double"
This reverts commit 5e516297e346f502cf2a47a74a72714bd5448209.

+++ cppcore/include/detail/strategy.hpp
@@ -12,8 +12,8 @@ namespace cpb { namespace detail {
  */
 template<class BaseStrategy, template<class> class Strategy>
 class MakeStrategy {
-    static_assert(std::is_base_of<BaseStrategy, Strategy<double>>::value, "");
-    using Config = typename Strategy<double>::Config;
+    static_assert(std::is_base_of<BaseStrategy, Strategy<float>>::value, "");
+    using Config = typename Strategy<float>::Config;
     Config config;
 
 public:
